Added tests for the parser.
I need to refactor the tokenizer. When a movement
token is encountered we should not be returning
a token that has a type and an attribute.

The way we are doing it now assumes we won't have
input like "50F 13L" and our program will break if
we try something like that.

Basically we need to keep movement tokens and the
associated number as separate entities.